+++ pkg/client/message/message_test.go
@@ -10,13 +10,13 @@ import (
 )
 
 const (
-	tcSalt    = "1e0803726723f2ec609a048aa7779f772ebf7002489d2e2f7bc6ae91e8eb619aba74e794c2d0a00e16bda23d75076513"
-	tcSession = "b53ac6435763f917e6da7cce7d7d37dc658de791f2da58655f1a715ac44d0b55c22342fa1287b6bc0bfadd4ab7ff480dbdfd09d62e861073ffb9291a6b8835499ee290792015faba1ef21c51facbe09c0c0c7454eb132a4219b6781a5673af864649cb8f8054b0744027fb5ba28894864935a45e133677fe49e8e92de42963f4"
-	tcSender  = "3ec534f2fa1df35376c55256bb1ee02e8a808d1813a9f4cac7537c232a2cb481987cfab744a65b75e8c707674e9996ef354a5616dd8a194337c2647815a0cbd92514529d32719a7e2b97ee52fe390e7653c88ff652ec53cddd2d6e3aac40c2adffbaf4758b406cfc3ee943c780ada11a2dc602c9fbe02821ce9428150c81f283c42a7f30eeb5ccee05375a5a93c0cbc10c8a3edac448385e0632c5dc"
+	tcSalt    = "f82291dde6c82aefad9c56462a5537a488be28a9decc1b5002c71e585a678283bad09298134d83bed2ad9fdf0a84a549"
+	tcSession = "0eaf199a0458c25ba7433b06a1ca58cb9938e3a29208aff0a1aa7e7b52beced29cce2e309e89d09321bdee9feea9b084bf9aab4e361a6d15f2d502e451271ea8e2e87c6da329c9019e42356fb0e50afb47aacb15e0586c3d1a678d579f82e636ecb6cb3e8b8bcc082621e10952a5f4619d539bbcb9639a5ad546549490a0f9de"
+	tcSender  = "ced82379a18f3fe84d839f8b4d8265c671a861b5140dc9021398c621d95c60365f8e7b19cdc47af1c0a92aabfc80c98731af5de901dbedb5d6d52b55c8bb3a7dab411a46891fb07f4308374139d55da01a952d637f00d8e3fbb457e504d6724e8b0b33563b0f26998434907eb6818285fadd39603ec8a087f30923ca4169b1fc8ed60fbd74579fe53693e357ce79717fcb32a8cb9a5f5dc48bf37b18"
 
-	tcHash  = "cbf42e03ea0760e86248ce59e822b4abf47239e86cd24d371dd87adf64232bac"
-	tcSign  = "2c2ebfe79d3911b9e987b1e27821dc834fb36352e3ac11dc2f590a6163d8f054dd308f84607337820ea3fdca4587c56f83b80d9578d2750ecacb495aae994a37c1ba66d17fcb4b87898ba4e256ffcdc629e68c459c48c5963f61b180ae47d4d44647df08a1ed2ea9458ff2f5400a7ec8121be37835d5e535f61517a987aba8b1fe52703aa7267e239e2a726a2dbf4935"
-	tcProof = 1651
+	tcSign  = "0af1958672e6d7a3530588bacf87fa1191d9a075cd38d5651f84a5f4dc97c7605e15e0170fcf27469c29a0c8d6d61b94aeeab01e6caf5d012c3d6ab5eeec2c7d88252fc41db0fd2c27484f37fcb5af8a868b2c0a34771fe667b66a3fd87cca17fe7f0e979861fdfcfa83483bd1a5776a9b10a43c61e91a9a810af30c8fee6391f2005694e109898bbd68f65c7089e934"
+	tcHash  = "f74311297112ca6ddfa2602e5c2b8a9d474f944987f71df764d3620b0ee3aa05"
+	tcProof = 332
 )
 
 func TestMessage(t *testing.T) {
@@ -27,8 +27,8 @@ func TestMessage(t *testing.T) {
 	}
 
 	params := NewSettings(&SSettings{
+		FMessageSizeBytes: (2 << 10),
 		FWorkSizeBits:     testutils.TCWorkSize,
-		FMessageSizeBytes: testutils.TCMessageSize,
 	})
 	msg := LoadMessage(params, msgBytes)
 	if msg == nil {
@@ -51,13 +51,13 @@ func TestMessage(t *testing.T) {
 		return
 	}
 
-	if !bytes.Equal(msg.GetBody().GetHash(), encoding.HexDecode(tcHash)) {
-		t.Error("incorrect hash value")
+	if !bytes.Equal(msg.GetBody().GetSign(), encoding.HexDecode(tcSign)) {
+		t.Error("incorrect sign value")
 		return
 	}
 
-	if !bytes.Equal(msg.GetBody().GetSign(), encoding.HexDecode(tcSign)) {
-		t.Error("incorrect sign value")
+	if !bytes.Equal(msg.GetBody().GetHash(), encoding.HexDecode(tcHash)) {
+		t.Error("incorrect hash value")
 		return
 	}